Abstracted a lot of Record class functionality into a MappingDB class.  Created Network class to represent Infoblox Network
