+++ infoblox/__init__.py
@@ -6,3 +6,4 @@ from infoblox.record import Host
 from infoblox.record import HostIPv4
 from infoblox.record import HostIPv6
 from infoblox.record import IPv4Address
+from infoblox.network import Network
